refactored ig name to snapshot
