+++ MapboxAndroidDemo/src/main/res/layout/activity_snapshot_notification.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".examples.ig.SnapshotNotificationActivity">
+    tools:context=".examples.snapshot.SnapshotNotificationActivity">
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
         android:layout_width="match_parent"
